refactor test: moved from #show to #edit_spec
